Accept fractional position for balance trade

+++ sysproduction/interactive_order_stack.py
@@ -184,7 +184,7 @@ def create_balance_trade(data):
         data)
     fill_qty = get_and_convert(
         "Quantity ",
-        type_expected=int,
+        type_expected=float,
         allow_default=False)
     filled_price = get_and_convert(
         "Filled price", type_expected=float, allow_default=False
